Fix a bug
WIP. We still need to test if h(x) is correct and does not overflow again

+++ src/headers/manhattan-hashing.hpp
@@ -15,6 +15,6 @@ extern unsigned long M;
 // uniform_real_distribution components
 float* calculateURDComponents(uint64_t);
 int* calculateA_IComponents(unsigned char *, float *, uint64_t);
-int calculateH_XComponents(int*, uint64_t);
-unsigned int exponentiationModulo(unsigned int , unsigned int , unsigned long);
+unsigned long int calculateH_XComponents(int*, uint64_t);
+unsigned long long int exponentiationModulo(unsigned int , unsigned int , unsigned long);
 unsigned long customModulo(unsigned long , unsigned long);
